Move LessonList into its own subdir
